+++ stockosaurus/stockosaurus/urls.py
@@ -5,6 +5,7 @@ from django.urls import path, include
 
 urlpatterns = [
     path('', include('stocks.urls')),
+    path('', include('users.urls')),
     path('admin/', admin.site.urls),
     path('api-auth/', include('rest_framework.urls')),
