+++ stockosaurus/stocks/views.py
@@ -1,11 +1,10 @@
-from rest_framework import generics, permissions, renderers
-from django.contrib.auth.models import User
+from rest_framework import generics, permissions
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
 from rest_framework.reverse import reverse
 
 from .models import StockPrice
-from .serializers import StockPriceSerializer, UserSerializer
+from .serializers import StockPriceSerializer
 from .permissions import IsOwnerOrReadOnly
 
 
@@ -44,12 +43,3 @@ class StockDetail(generics.RetrieveUpdateDestroyAPIView):
 	serializer_class = StockPriceSerializer
 	permission_classes = [permissions.IsAuthenticatedOrReadOnly, IsOwnerOrReadOnly]
 
-
-class UserList(generics.ListAPIView):
-	queryset = User.objects.all()
-	serializer_class = UserSerializer
-
-
-class UserDetail(generics.RetrieveAPIView):
-	queryset = User.objects.all()
-	serializer_class = UserSerializer
