+++ stockosaurus/users/models.py
@@ -1,0 +1,3 @@
+from django.db import models
+
+owner = models.ForeignKey('auth.User', related_name='snippets', on_delete=models.CASCADE)
