+++ stockosaurus/users/serializers.py
@@ -1,0 +1,10 @@
+from rest_framework import serializers
+from django.contrib.auth.models import User
+
+
+class UserSerializer(serializers.HyperlinkedModelSerializer):
+    stocks = serializers.HyperlinkedRelatedField(many=True, view_name='user-detail', read_only=True)
+
+    class Meta:
+        model = User
+        fields = ['id', 'username', 'stocks']
