+++ stockosaurus/users/views.py
@@ -1,0 +1,13 @@
+from rest_framework import generics
+from django.contrib.auth.models import User
+from .serializers import UserSerializer
+
+
+class UserList(generics.ListAPIView):
+    queryset = User.objects.all()
+    serializer_class = UserSerializer
+
+
+class UserDetail(generics.RetrieveAPIView):
+    queryset = User.objects.all()
+    serializer_class = UserSerializer
